OTT-499: Incorporating review comments
