+++ exchange/utils.go
@@ -159,225 +159,6 @@ func cleanOpenRTBRequests(ctx context.Context,
 	return
 }
 
-// updateContentObjectForBidder updates the content object for each bidder based on content transparency rules
-func updateContentObjectForBidder(allBidderRequests []BidderRequest, requestExt *openrtb_ext.ExtRequest) {
-
-	if requestExt != nil && requestExt.Prebid.Transparency != nil {
-
-		var contentObject *openrtb2.Content
-		isApp := false
-		bidderRequest := allBidderRequests[0]
-		if bidderRequest.BidRequest.App != nil && bidderRequest.BidRequest.App.Content != nil {
-			contentObject = bidderRequest.BidRequest.App.Content
-			isApp = true
-		} else if bidderRequest.BidRequest.Site != nil && bidderRequest.BidRequest.Site.Content != nil {
-			contentObject = bidderRequest.BidRequest.Site.Content
-		} else {
-			return
-		}
-
-		var (
-			rule openrtb_ext.TransparencyRule
-			ok   bool
-		)
-
-		rules, defaultRule := getTransparencyRules(requestExt.Prebid.Transparency)
-
-		for i, bidderRequest := range allBidderRequests {
-
-			newContentObject := &openrtb2.Content{}
-
-			if len(rules) == 0 {
-				newContentObject = nil
-			} else {
-
-				if rule, ok = rules[string(bidderRequest.BidderName)]; !ok {
-					rule = defaultRule
-				}
-
-				if len(rule.Keys) == 0 {
-					if rule.Include {
-						*newContentObject = *contentObject
-					} else {
-						newContentObject = nil
-					}
-				} else {
-					createNewContentObject(newContentObject, contentObject, rule.Include, rule.Keys)
-				}
-			}
-
-			bidderRequest.BidRequest = copyRequest(bidderRequest.BidRequest, newContentObject, isApp)
-			allBidderRequests[i] = bidderRequest
-		}
-	}
-}
-
-func copyRequest(originalRequest *openrtb2.BidRequest, contentObject *openrtb2.Content, isApp bool) *openrtb2.BidRequest {
-
-	newRequest := *originalRequest
-
-	if isApp {
-		app := *originalRequest.App
-		app.Content = contentObject
-		newRequest.App = &app
-	} else {
-		site := *originalRequest.Site
-		site.Content = contentObject
-		newRequest.Site = &site
-	}
-	return &newRequest
-}
-
-func createNewContentObject(newContentObject, contentObject *openrtb2.Content, include bool, keys []string) {
-
-	if !include {
-		*newContentObject = *contentObject
-		for _, key := range keys {
-
-			switch key {
-			case "id":
-				newContentObject.ID = ""
-			case "episode":
-				newContentObject.Episode = 0
-			case "title":
-				newContentObject.Title = ""
-			case "series":
-				newContentObject.Series = ""
-			case "season":
-				newContentObject.Season = ""
-			case "artist":
-				newContentObject.Artist = ""
-			case "genre":
-				newContentObject.Genre = ""
-			case "album":
-				newContentObject.Album = ""
-			case "isrc":
-				newContentObject.ISRC = ""
-			case "producer":
-				newContentObject.Producer = nil
-			case "url":
-				newContentObject.URL = ""
-			case "cat":
-				newContentObject.Cat = nil
-			case "prodq":
-				newContentObject.ProdQ = nil
-			case "videoquality":
-				newContentObject.VideoQuality = nil
-			case "context":
-				newContentObject.Context = 0
-			case "contentrating":
-				newContentObject.ContentRating = ""
-			case "userrating":
-				newContentObject.UserRating = ""
-			case "qagmediarating":
-				newContentObject.QAGMediaRating = 0
-			case "keywords":
-				newContentObject.Keywords = ""
-			case "livestream":
-				newContentObject.LiveStream = 0
-			case "sourcerelationship":
-				newContentObject.SourceRelationship = 0
-			case "len":
-				newContentObject.Len = 0
-			case "language":
-				newContentObject.Language = ""
-			case "embeddable":
-				newContentObject.Embeddable = 0
-			case "data":
-				newContentObject.Data = nil
-			case "ext":
-				newContentObject.Ext = nil
-			}
-		}
-		return
-	}
-
-	for _, key := range keys {
-		switch key {
-		case "id":
-			newContentObject.ID = contentObject.ID
-		case "episode":
-			newContentObject.Episode = contentObject.Episode
-		case "title":
-			newContentObject.Title = contentObject.Title
-		case "series":
-			newContentObject.Series = contentObject.Series
-		case "season":
-			newContentObject.Season = contentObject.Season
-		case "artist":
-			newContentObject.Artist = contentObject.Artist
-		case "genre":
-			newContentObject.Genre = contentObject.Genre
-		case "album":
-			newContentObject.Album = contentObject.Album
-		case "isrc":
-			newContentObject.ISRC = contentObject.ISRC
-		case "producer":
-			if contentObject.Producer != nil {
-				producer := *contentObject.Producer
-				newContentObject.Producer = &producer
-			}
-		case "url":
-			newContentObject.URL = contentObject.URL
-		case "cat":
-			newContentObject.Cat = contentObject.Cat
-		case "prodq":
-			if contentObject.ProdQ != nil {
-				prodQ := *contentObject.ProdQ
-				newContentObject.ProdQ = &prodQ
-			}
-		case "videoquality":
-			if contentObject.VideoQuality != nil {
-				videoQuality := *contentObject.VideoQuality
-				newContentObject.VideoQuality = &videoQuality
-			}
-		case "context":
-			newContentObject.Context = contentObject.Context
-		case "contentrating":
-			newContentObject.ContentRating = contentObject.ContentRating
-		case "userrating":
-			newContentObject.UserRating = contentObject.UserRating
-		case "qagmediarating":
-			newContentObject.QAGMediaRating = contentObject.QAGMediaRating
-		case "keywords":
-			newContentObject.Keywords = contentObject.Keywords
-		case "livestream":
-			newContentObject.LiveStream = contentObject.LiveStream
-		case "sourcerelationship":
-			newContentObject.SourceRelationship = contentObject.SourceRelationship
-		case "len":
-			newContentObject.Len = contentObject.Len
-		case "language":
-			newContentObject.Language = contentObject.Language
-		case "embeddable":
-			newContentObject.Embeddable = contentObject.Embeddable
-		case "data":
-			if contentObject.Data != nil {
-				newContentObject.Data = contentObject.Data
-			}
-		case "ext":
-			newContentObject.Ext = contentObject.Ext
-		}
-	}
-}
-
-//  getTransparencyRules return content object transparency rule for a bidder and default rule
-func getTransparencyRules(transparencyExt *openrtb_ext.TransparencyExt) (map[string]openrtb_ext.TransparencyRule, openrtb_ext.TransparencyRule) {
-
-	rules := transparencyExt.Content
-
-	// Dont send content object if no rule and default is not present
-	defaultRule := openrtb_ext.TransparencyRule{
-		Include: false,
-	}
-
-	if rule, ok := rules["default"]; ok {
-		defaultRule = rule
-	}
-
-	return rules, defaultRule
-}
-
 func gdprEnabled(account *config.Account, privacyConfig config.Privacy, integrationType config.IntegrationType) bool {
 	if accountEnabled := account.GDPR.EnabledForIntegrationType(integrationType); accountEnabled != nil {
 		return *accountEnabled
